requirements.txt
et-xmlfile==1.1.0
mysql-connector-python==9.0.0
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
